% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apriori_ARM_tool.R
\name{Abstract_ARM}
\alias{Abstract_ARM}
\title{Perform association-rules mining on lemmatized abstracts.}
\arguments{
\item{data}{Your text_parser() results go here.}

\item{min_supp}{Define the minimum support threshold for the rules (e.g 0.01 or 0.10)}

\item{min_conf}{Define the minimum confidence threshold for the rules (e.g 0.50 or 0.90)}

\item{min_p}{Define the minimum p-value threshold for the rules (e.g 0.05 or 0.0.005)}
}
\description{
Using the results from the Text_Parser() function, generate a dataframe of association-rules from the corpus. P-values for results are calculated using Fischer's exact test, and FDR are used to remove false rejections of the null hypothesis.
Be aware of your ARM mining parameters because these can seriously impact your computational load.
}
\examples{
Abstract_ARM(data <- data_from_Text_Parser, min_supp = 0.01, min_conf = 0.75, min_p = 0.005)
}
\keyword{Processing}
\keyword{algorithm,}
\keyword{arules,}
\keyword{association-rules}
\keyword{mining,}
\keyword{statistics,}
