% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{Abstract_Trimmer}
\alias{Abstract_Trimmer}
\title{An abstract text trimming function}
\arguments{
\item{raw_data}{Input your entire corpus of abstracts. (typically obtained from the info_retrieval function)}

\item{fraction}{The decimal represenitng the fraction of each abstract to trim. We recommend (0.2) or 20\%}
}
\description{
You can remove any \% of text from an abstract corpus using this function. It also removes empty abstracts.The output is trimmed and non "NA" abstract data.
}
\examples{
Abstract_Trimmer(raw_data = info_retrieval_results, fraction = 0.2)
}
\keyword{NLP,}
\keyword{Text}
\keyword{language}
\keyword{natural}
\keyword{processing}
\keyword{reduction,}
