% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_Retrieval.R
\name{europepmc_retrieval}
\alias{europepmc_retrieval}
\title{The Europe PMC search and retrieve function}
\usage{
europepmc_retrieval(query, retmax)
}
\arguments{
\item{retmax}{Define how many entries you'd like to access for a query. Keep it between 500-1000 for fastest results.}

\item{Query}{Your typical PubMed query. Optimizing your query using the proper pubmed syntax improves results!}
}
\description{
This function allows you to query and access key Pubmed entry data from one of two databases: PubMed and EuropePMC.
}
\examples{
europepmc_retrieval(Query = "Vape smoking AND toxicity", retmax = 750)
}
\keyword{EuropePMC,}
\keyword{PMC,}
\keyword{Pubmed,}
\keyword{Query,}
\keyword{Retrieval}
