% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_Retrieval.R
\name{info_retrieval}
\alias{info_retrieval}
\title{Extract PubMed or Europe PMC entry information for a query}
\arguments{
\item{query}{Your typical PubMed query. Optimizing your query using the proper PubMed or Europe PMC syntax improves results.}

\item{how_many_articles}{An integer the number of entries to retrieve. PubMed is limited to about 1500 articles, which Europe PMC can pull around 20,000.}

\item{database}{Define the PubMed database to retrieve from: "pubmed" or "pmc".}
}
\description{
Combines functionality from two R-packages: "easyPubMed" and "europepmc" for more reliable PubMed abstract extraction.
}
\examples{
info_retireval(query = "Vape smoking AND toxicity", how_many_articles = 750, database = "pubmed")
}
\keyword{Article}
\keyword{Europe}
\keyword{PMC,}
\keyword{PubMed,}
\keyword{Retrieval,}
