% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KAF_Deduplication.R
\name{KAF_Deduplication}
\alias{KAF_Deduplication}
\title{Removes redundant rules prior to matching rules to AOP-wiki}
\usage{
KAF_Deduplication(rules)
}
\arguments{
\item{rules}{Rules generated through easyKAF or the manual KAF pipeline.}
}
\description{
This function returns a smaller ruleset containing less near-duplicates using Locality Sensitive Hashing (LSH).
}
\examples{
KAF_Deduplication(rules)
}
\keyword{Hashing,}
\keyword{LSH,}
\keyword{Locality}
\keyword{deduplication}
\keyword{hashing,}
\keyword{sensitive}
