% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOP_Classification_Study.R
\name{KAFxAOP}
\alias{KAFxAOP}
\title{Uses Locality Sensitive Hashing and Jaccard Similarity to evaluate}
\usage{
KAFxAOP(sample)
}
\arguments{
\item{rules}{generated through}
}
\description{
This function returns a smaller ruleset containing less near-duplicates using Locality Sensitive Hashing (LSH).
}
\examples{
KAFxAOP(rules)
}
\keyword{AOP}
\keyword{Hashing,}
\keyword{Jaccard}
\keyword{LSH,}
\keyword{Locality}
\keyword{Similarity,}
\keyword{deduplication,}
\keyword{fuzzy}
\keyword{hashing,}
\keyword{matching,}
\keyword{sensitive}
