% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{Text_Parser}
\alias{Text_Parser}
\title{The one-and-done NLP function which handles pre-processing your retrieved abstract}
\usage{
Text_Parser(data, venv_path, lang_model, reduced_search)
}
\arguments{
\item{data}{Input your entire corpus of abstracts. (typically obtained from the info_retrieval function)}

\item{venv_path}{This is the path to your initialized python virtual environment. It must not contain any errors. If the venv does not initialize or the function does not function properly, force quit R studio and re-write the venv path until it works.}

\item{lang_model}{This is the name of the spaCy language model you install (e.g en_core_web_sm)}

\item{reduced_search}{Identify the percentage of text to remove from the beginning of the abstract (We recommend 0.2)}
}
\description{
This trims abstracts, intializes a spaCy language model, parses them, removes sentences containing negations, and returns lemma of potentially significant keywords. IMPORTANT: This function requires an initialized python venv with spacy and a spacy english model installed. The path to this venv is crucial.
}
\examples{
Text_Parser(data = info_retrieval_fxn_output, venv_path = "your/exact/venv/path", lang_model = "en_core_web_sm", reduced_search = 0.2)
}
\keyword{NLP,}
\keyword{Text}
\keyword{language}
\keyword{modifier}
\keyword{natural}
\keyword{negation}
\keyword{parsing,}
\keyword{processing,}
\keyword{reduction,}
