% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_Retrieval.R
\name{info_retrieval}
\alias{info_retrieval}
\title{A fully integrated Pubmed abstract retrieval function}
\arguments{
\item{Query}{Your typical PubMed query. Optimizing your query using the proper pubmed syntax improves results!}

\item{how_many_articles}{Define how many entries you'd like to access for a query. Keep it between 500-1000 for fastest results.}

\item{database}{Define the database you'd like to use: "pubmed" or "pmc".}
}
\description{
This function allows you to query and access key Pubmed entry data from one of two databases: PubMed and EuropePMC.
}
\examples{
info_retireval(Query = "Vape smoking AND toxicity", how_many_articles = 750, database = "pubmed")
}
\keyword{PMC,}
\keyword{Pubmed,}
\keyword{Query}
